# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  backend_internal_httpresp.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  backend_internal_httpresp.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/backend_internal_httpresp.ErrorBody'
    type: object
...
  internal_approvals.ApprovalListResponse:
    properties:
      approvals:
        items:
          $ref: '#/definitions/internal_approvals.ApprovalResponse'
        type: array
      limit:
        example: 50
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 15
        type: integer
    type: object
  internal_approvals.ApprovalResponse:
    properties:
      application_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      cover_letter_preview:
        example: /storage/cover-letters/550e8400-e29b-41d4-a716-446655440001.pdf
        type: string
      created_at:
        example: "2026-06-19T10:00:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      job_snapshot:
        $ref: '#/definitions/internal_approvals.JobSnapshot'
      rejection_reason:
        example: Salary too low
        type: string
      resume_preview_path:
        example: /storage/resumes/550e8400-e29b-41d4-a716-446655440000.pdf
        type: string
      reviewed_at:
        example: "2026-06-20T14:30:00Z"
        type: string
      status:
        enum:
        - pending
        - approved
        - rejected
        example: pending
        type: string
    type: object
  internal_approvals.ApprovePartialResponse:
    properties:
      approval:
        $ref: '#/definitions/internal_approvals.ApprovalResponse'
      status:
        example: approved
        type: string
      warning:
        example: application submission queued for retry
        type: string
    type: object
  internal_approvals.ApproveRequest:
    type: object
  internal_approvals.JobSnapshot:
    properties:
      company:
        type: string
      description:
        type: string
      location:
        type: string
      requirements:
        items:
          type: string
        type: array
      score:
        type: number
      scored_at:
        type: string
      tier:
        description: '"review" — only review-tier jobs need approval'
        type: string
      title:
        type: string
      url:
        type: string
    type: object
  internal_approvals.RejectRequest:
    properties:
      reason:
        description: Reason is required for audit trail.
        example: Salary too low for current market
        type: string
    required:
    - reason
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/your-org/ai-job-search-agent/issues
  description: Local-first AI-powered job search automation platform. Discovers, scores,
    and applies to jobs with tailored resumes and cover letters.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/your-org/ai-job-search-agent
  title: AI Job Search Agent API
  version: "1.0"
paths:
...
  /approvals:
    get:
      consumes:
      - application/json
      description: Get paginated list of approval requests with optional filters
      parameters:
      - description: Filter by status
        enum:
        - pending
        - approved
        - rejected
        in: query
        name: status
        type: string
      - description: Filter by application UUID
        in: query
        name: application_id
        type: string
      - default: 50
        description: Results per page (max 100)
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated approval requests
          schema:
            $ref: '#/definitions/internal_approvals.ApprovalListResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List approval requests
      tags:
      - Approvals
  /approvals/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed approval request with job snapshot and previews
      parameters:
      - description: Approval UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Approval request details
          schema:
            $ref: '#/definitions/internal_approvals.ApprovalResponse'
        "400":
          description: Invalid approval ID
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "404":
          description: Approval request not found
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get approval request by ID
      tags:
      - Approvals
  /approvals/{id}/approve:
    post:
      consumes:
      - application/json
      description: Approve the application and trigger automatic submission workflow
      parameters:
      - description: Approval UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Empty body - approval decision is in endpoint
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_approvals.ApproveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Application approved and submission queued
          schema:
            $ref: '#/definitions/internal_approvals.ApprovalResponse'
        "207":
          description: Approved but submission dispatch failed (queued for retry)
          schema:
            $ref: '#/definitions/internal_approvals.ApprovePartialResponse'
        "400":
          description: Invalid approval ID
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "404":
          description: Approval not found
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "409":
          description: Invalid status transition
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Approve application
      tags:
      - Approvals
  /approvals/{id}/reject:
    post:
      consumes:
      - application/json
      description: Reject the application with a required reason
      parameters:
      - description: Approval UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Rejection reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_approvals.RejectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Application rejected
          schema:
            $ref: '#/definitions/internal_approvals.ApprovalResponse'
        "400":
          description: Invalid approval ID or missing reason
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "404":
          description: Approval not found
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "409":
          description: Invalid status transition
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reject application
      tags:
      - Approvals
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization:
      Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...